% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_fun}
\alias{n_fun}
\title{Build a median/size label for plot annotations}
\usage{
n_fun(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A data frame with the median and label.
}
\description{
Build a median/size label for plot annotations
}
