% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.zero}
\alias{na.zero}
\title{Replace NA values with zero}
\usage{
na.zero(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector with NAs replaced by zeros.
}
\description{
Replace NA values with zero
}
\examples{
na.zero(c(NA, 1, NA, 2))
}
