% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Negate \code{\%in\%} membership}
\usage{
\%!in\%(x, y)
}
\arguments{
\item{x}{Vector of values to test.}

\item{y}{Vector of values to match against.}
}
\value{
Logical vector indicating non-membership.
}
\description{
Negate \code{\%in\%} membership.
}
