% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pathPrep}
\alias{pathPrep}
\title{Convert Windows paths to R-friendly format}
\usage{
pathPrep(path = "clipboard", read_fn = NULL, write_fn = NULL)
}
\arguments{
\item{path}{Path to convert or the string "clipboard" to read from the clipboard.}

\item{read_fn}{Optional custom function to read from the clipboard.}

\item{write_fn}{Optional custom function to write to the clipboard.}
}
\value{
A normalized path string.
}
\description{
Convert Windows paths to R-friendly format
}
