% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R, R/zzz-aliases.R
\name{ggwithinstatsWithPriorNormalityCheckAsterisk}
\alias{ggwithinstatsWithPriorNormalityCheckAsterisk}
\alias{plot_within_stats_asterisk}
\title{Check the data's distribution. If non-normal, take the non-parametric variant of \emph{ggwithinstats}.
x and y have to be in parentheses, e.g., "ConditionID". Add Asterisks instead of p-values.}
\usage{
ggwithinstatsWithPriorNormalityCheckAsterisk(
  data,
  x,
  y,
  ylab,
  xlabels,
  plotType = "boxviolin"
)

plot_within_stats_asterisk(data, x, y, ylab, xlabels, plotType = "boxviolin")
}
\arguments{
\item{data}{the data frame}

\item{x}{the independent variable, most likely "ConditionID"}

\item{y}{the dependent variable under investigation}

\item{ylab}{label to be shown for the dependent variable}

\item{xlabels}{labels to be used for the x-axis}

\item{plotType}{either "box", "violin", or "boxviolin" (default)}
}
\value{
A \code{ggplot} object produced by \code{ggstatsplot::ggwithinstats}
with additional significance annotations, which can be printed or modified.
}
\description{
Check the data's distribution. If non-normal, take the non-parametric variant of \emph{ggwithinstats}.
x and y have to be in parentheses, e.g., "ConditionID". Add Asterisks instead of p-values.
}
\examples{
\donttest{

set.seed(123)

# Toy within-subject style data
main_df <- data.frame(
  Participant = factor(rep(1:20, each = 3)),
  CondID      = factor(rep(c("A", "B", "C"), times = 20)),
  tlx_mental  = rnorm(60, mean = 50, sd = 10)
)

# Custom x-axis labels
labels_xlab <- c("Condition A", "Condition B", "Condition C")


ggwithinstatsWithPriorNormalityCheckAsterisk(
  data = main_df,
  x = "CondID", y = "tlx_mental",
  ylab = "Mental Demand", xlabels = labels_xlab
)
}
}
