% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_in}
\alias{not_in}
\alias{\%!in\%}
\title{Negate \code{\%in\%} membership}
\usage{
not_in(x, y)

x \%!in\% y
}
\arguments{
\item{x}{Vector of values to test.}

\item{y}{Vector of values to match against.}
}
\value{
Logical vector indicating non-membership.
}
\description{
Negate \code{\%in\%} membership
}
