% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R, R/zzz-aliases.R
\name{reportMeanAndSD}
\alias{reportMeanAndSD}
\alias{report_mean_sd}
\title{Report the mean and standard deviation of a dependent variable for all levels of an independent variable rounded to the 2nd digit.}
\usage{
reportMeanAndSD(data, iv = "testiv", dv = "testdv", sink_to = NULL)

report_mean_sd(data, iv = "testiv", dv = "testdv", sink_to = NULL)
}
\arguments{
\item{data}{the data frame}

\item{iv}{the independent variable}

\item{dv}{the dependent variable}

\item{sink_to}{optional path of a \code{.tex} file to write the lines to,
so a manuscript can \code{\\input{}} them}
}
\value{
Invisibly returns the formatted lines as a character vector;
the text is also emitted via \code{message()}.
}
\description{
#' To easily copy and paste the results to your manuscript, the following commands must be defined in Latex:
\code{\\newcommand{\\m}{\\textit{M=}}}
\code{\\newcommand{\\sd}{\\textit{SD=}}}
}
\examples{
\donttest{

example_data <- data.frame(Condition = rep(c("A", "B", "C"),
each = 10), TLX1 = stats::rnorm(30))

reportMeanAndSD(example_data, iv = "Condition", dv = "TLX1")
}
}
