% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{reportNPAV}
\alias{reportNPAV}
\title{Generate the Latex-text based on the NPAV by Lüpsen (see \url{https://www.uni-koeln.de/~luepsen/R/}).
Only significant main and interaction effects are reported.
P-values are rounded for the third digit and partial eta squared values are provided when possible.
Attention: the independent variables of the formula and the term specifying the participant must be factors (i.e., use as.factor()).}
\usage{
reportNPAV(
  model,
  dv = "Testdependentvariable",
  write_to_clipboard = FALSE,
  sink_to = NULL
)
}
\arguments{
\item{model}{the model of the np.anova}

\item{dv}{the name of the dependent variable that should be reported}

\item{write_to_clipboard}{whether to write to the clipboard}

\item{sink_to}{optional path of a \code{.tex} file to write the sentences to,
so a manuscript can \code{\\input{}} them}
}
\value{
Invisibly returns the reported sentence(s) as a character vector;
the text is also emitted via \code{message()}.
}
\description{
Deprecated: \code{reportNPAV()} will be removed in a future release.
Use \code{reportART()} with ARTool instead.
}
\details{
To easily copy and paste the results to your manuscript, the following commands must be defined in Latex:
\code{\\newcommand{\\F}[3]{$F({#1},{#2})={#3}$}}
\code{\\newcommand{\\p}{\\textit{p=}}}
\code{\\newcommand{\\pminor}{\\textit{p$<$}}}
}
\examples{
model <- data.frame(
  Df = c(1, 1, 10),
  `F value` = c(6.12, 5.01, NA),
  `Pr(>F)` = c(0.033, 0.045, NA),
  check.names = FALSE
)
rownames(model) <- c("Video", "gesture:eHMI", "Residuals")
reportNPAV(model, dv = "mental workload")
}
